+++ bevy_replicon_example_backend/src/server.rs
@@ -20,38 +20,37 @@ impl Plugin for RepliconExampleServerPlugin {
         app.add_systems(
             PreUpdate,
             (
+                (
+                    receive_packets.run_if(resource_exists::<ExampleServer>),
+                    // Run after since the resource might be removed after receiving packets.
+                    set_stopped.run_if(resource_removed::<ExampleServer>),
+                )
+                    .chain(),
                 set_running.run_if(resource_added::<ExampleServer>),
-                receive_packets.run_if(resource_exists::<ExampleServer>),
             )
-                .chain()
                 .in_set(ServerSet::ReceivePackets),
         )
         .add_systems(
             PostUpdate,
-            (
-                set_stopped
-                    .in_set(ServerSet::PrepareSend)
-                    .run_if(resource_removed::<ExampleServer>),
-                send_packets
-                    .in_set(ServerSet::SendPackets)
-                    .run_if(resource_exists::<ExampleServer>),
-            ),
+            send_packets
+                .run_if(resource_exists::<ExampleServer>)
+                .in_set(ServerSet::SendPackets),
         );
     }
 }
 
-fn set_stopped(mut server: ResMut<RepliconServer>) {
-    server.set_running(false);
+fn set_running(mut state: ResMut<NextState<ServerState>>) {
+    state.set(ServerState::Running);
 }
 
-fn set_running(mut server: ResMut<RepliconServer>) {
-    server.set_running(true);
+fn set_stopped(mut state: ResMut<NextState<ServerState>>) {
+    state.set(ServerState::Stopped);
 }
 
 fn receive_packets(
     mut commands: Commands,
+    mut messages: ResMut<ServerMessages>,
     server: Res<ExampleServer>,
-    mut replicon_server: ResMut<RepliconServer>,
     mut clients: Query<(Entity, &mut ExampleConnection, Option<&ConditionerConfig>)>,
     global_config: Option<Res<ConditionerConfig>>,
 ) {
@@ -117,7 +116,7 @@ fn receive_packets(
         }
 
         while let Some((channel_id, message)) = connection.conditioner.pop(now) {
-            replicon_server.insert_received(client, channel_id, message)
+            messages.insert_received(client, channel_id, message)
         }
     }
 }
@@ -125,10 +124,10 @@ fn receive_packets(
 fn send_packets(
     mut commands: Commands,
     mut disconnect_events: EventReader<DisconnectRequest>,
-    mut replicon_server: ResMut<RepliconServer>,
+    mut messages: ResMut<ServerMessages>,
     mut clients: Query<&mut ExampleConnection>,
 ) {
-    for (client, channel_id, message) in replicon_server.drain_sent() {
+    for (client, channel_id, message) in messages.drain_sent() {
         let mut connection = clients
             .get_mut(client)
             .expect("all connected clients should have streams");