+++ src/test_app.rs
@@ -8,7 +8,7 @@ Extension for [`App`] to communicate with other instances like it's a server.
 # Example
 
 ```
-use bevy::prelude::*;
+use bevy::{prelude::*, state::app::StatesPlugin};
 use bevy_replicon::{prelude::*, test_app::ServerTestAppExt};
 
 let mut server_app = App::new();
@@ -16,6 +16,7 @@ let mut client_app = App::new();
 for app in [&mut server_app, &mut client_app] {
     app.add_plugins((
         MinimalPlugins,
+        StatesPlugin,
         // No messaging library plugin required.
         RepliconPlugins.set(ServerPlugin {
             tick_policy: TickPolicy::EveryFrame, // To tick each app update.
@@ -83,19 +84,26 @@ pub trait ServerTestAppExt {
 
 impl ServerTestAppExt for App {
     fn connect_client(&mut self, client_app: &mut App) {
-        let mut server = self.world_mut().resource_mut::<RepliconServer>();
-        server.set_running(true);
+        self.world_mut()
+            .resource_mut::<NextState<ServerState>>()
+            .set(ServerState::Running);
+
         let client_entity = self
             .world_mut()
             .spawn(ConnectedClient { max_size: 1200 })
             .id();
 
-        let mut client = client_app.world_mut().resource_mut::<RepliconClient>();
-        assert!(
-            client.is_disconnected(),
+        assert_eq!(
+            *client_app.world_mut().resource::<State<ClientState>>(),
+            ClientState::Disconnected,
             "client can't be connected multiple times"
         );
-        client.set_status(RepliconClientStatus::Connected);
+
+        client_app
+            .world_mut()
+            .resource_mut::<NextState<ClientState>>()
+            .set(ClientState::Connected);
+
         client_app
             .world_mut()
             .insert_resource(TestClientEntity(client_entity));
@@ -109,8 +117,11 @@ impl ServerTestAppExt for App {
     }
 
     fn disconnect_client(&mut self, client_app: &mut App) {
-        let mut client = client_app.world_mut().resource_mut::<RepliconClient>();
-        client.set_status(RepliconClientStatus::Disconnected);
+        client_app
+            .world_mut()
+            .resource_mut::<NextState<ClientState>>()
+            .set(ClientState::Disconnected);
+
         let client_entity = *client_app
             .world_mut()
             .remove_resource::<TestClientEntity>()
@@ -124,16 +135,16 @@ impl ServerTestAppExt for App {
 
     fn exchange_with_client(&mut self, client_app: &mut App) {
         let client_entity = **client_app.world().resource::<TestClientEntity>();
-        let mut client = client_app.world_mut().resource_mut::<RepliconClient>();
+        let mut client_messages = client_app.world_mut().resource_mut::<ClientMessages>();
 
-        let mut server = self.world_mut().resource_mut::<RepliconServer>();
-        for (channel_id, message) in client.drain_sent() {
-            server.insert_received(client_entity, channel_id, message)
+        let mut server_messages = self.world_mut().resource_mut::<ServerMessages>();
+        for (channel_id, message) in client_messages.drain_sent() {
+            server_messages.insert_received(client_entity, channel_id, message)
         }
 
-        server.retain_sent(|(entity, channel_id, message)| {
+        server_messages.retain_sent(|(entity, channel_id, message)| {
             if *entity == client_entity {
-                client.insert_received(*channel_id, message.clone());
+                client_messages.insert_received(*channel_id, message.clone());
                 false
             } else {
                 true